Show/hide form on button click
Add status property to data. Add method to change status. Add v-if directive to show add card button if status is default. Add v-else directive to cancel to show if status is edit. Add v-if directive to hide form if status is default. Change input for message to text-area. Add basic styling for form inputs and textarea.